types organise and tidy
