+++ packages/engine/src/store/actionCreators/createReconcileSketchNodes.ts
@@ -26,11 +26,6 @@ export const createReconcileSketchNodes: SetterCreator<'reconcileSketchNodes'> =
       for (const nodeConfig of config.nodes) {
         // Find existing node for this key, if any.
 
-        // FIXME: There should never be a custom node as part of a sketch, types should know this
-        if (nodeConfig.nodeType === 'custom') {
-          continue
-        }
-
         let nodeId = findNodeWithKeyFromIdList(
           state.nodes,
           nodeConfig.key,
