+++ packages/engine/src/store/selectors/getNextEnumValue.ts
@@ -1,4 +1,4 @@
-import { EngineState, NodeEnumValue, ParamEnum } from '@store/types'
+import { EngineState, ParamEnum, ParamEnumValue } from '@store/types'
 
 export const getNextEnumValue = (targetNodeId: string) => (state: EngineState) => {
   const paramVal = state.paramValues[targetNodeId]
@@ -9,7 +9,7 @@ export const getNextEnumValue = (targetNodeId: string) => (state: EngineState) =
     return paramVal
   }
 
-  const options = param.options as { value: NodeEnumValue }[]
+  const options = param.options as { value: ParamEnumValue }[]
   const currentIndex = options.findIndex((option) => option.value === paramVal)
 
   if (currentIndex === -1) {
