+++ packages/engine/src/store/types/CustomNode.ts
@@ -1,0 +1,16 @@
+import { NodeBase, ConfigNodeBase } from './NodeBase'
+
+export type CustomNode = NodeBase & {
+  nodeType: 'custom'
+  customNodeType: string
+}
+
+export type ConfigCustomNode = ConfigNodeBase & {
+  nodeType: 'custom'
+  customNodeType: string
+}
+
+export type ConfigCustomNodeImported = ConfigCustomNode & {
+  groupIndex: number
+  title: string
+}
