+++ packages/engine/src/store/types/Input.ts
@@ -1,0 +1,7 @@
+import { NodeBase } from './NodeBase'
+
+export interface Input extends NodeBase {
+  nodeType: 'input'
+  inputType: string
+  targetNodeId: string
+}
