+++ packages/engine/src/store/types/Shot.ts
@@ -1,0 +1,18 @@
+import { NodeBase, ConfigNodeBase } from './NodeBase'
+
+export type Shot = NodeBase & {
+  nodeType: 'shot'
+  key: string
+  hidden?: boolean
+  groupIndex: number
+}
+
+export interface ConfigShot extends ConfigNodeBase {
+  nodeType: 'shot'
+}
+
+export interface ConfigShotImported extends ConfigShot {
+  groupIndex: number
+  title: string
+  nodeType: 'shot'
+}
